updated test for targetAcquisition
